requirements.txt
simple_pid
opencv_python
numpy
pyserial
loguru